rename Filters component
